factor out chain related functions to fix cirtular dependency issue
